refactor: フォルダ名の変更・コメントの修正
