+++ app/Http/Controllers/API/PostController.php
@@ -5,7 +5,7 @@ namespace App\Http\Controllers\API;
 use App\Events\ThreadBrowsing\PostStored;
 use App\Exceptions\ThreadNotFoundException;
 use App\Http\Controllers\Controller;
-use App\Http\Requests\Post\IndexRequest;
+use App\Http\Requests\Posts\IndexRequest;
 use App\Http\Resources\PostResource;
 use App\Services\Tables\PostService;
 use App\Services\ThreadImageService;
@@ -27,7 +27,7 @@ class PostController extends Controller
     /**
      * 該当スレッドの書き込み一覧を取得する
      *
-     * @param Request $request アクセス時のパラメータなど
+     * @param IndexRequest $request アクセス時のパラメータなど
      * @return PostResource 成形済みの書き込み一覧
      */
     public function index(IndexRequest $request): PostResource
